% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregados.R
\docType{data}
\name{agregados}
\alias{agregados}
\title{API servicodados.ibge ... Todos agregados disponíveis}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 3 columns.
}
\source{
\href{https://servicodados.ibge.gov.br/api/docs/agregados?versao=3}{IBGE API}
}
\usage{
data(agregados)
}
\description{
Para utilização inicial, base com todos agregados disponíveis
}
\examples{
data(agregados)

}
\keyword{agregados}
\keyword{metadados}
