% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidrameta.R
\docType{data}
\name{sidrameta}
\alias{sidrameta}
\title{API servicodados.ibge ... Todos agregados disponíveis}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 21470 rows and 4 columns.
}
\source{
\href{https://servicodados.ibge.gov.br/api/docs/agregados?versao=3}{IBGE API}
}
\usage{
data(agregados)

data(sidrameta)
}
\description{
Para utilização inicial, base com todos agregados disponíveis
}
\examples{
data(sidrameta)

}
\keyword{agregados}
\keyword{metadados}
