% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_agr.R
\name{tab_agr}
\alias{tab_agr}
\title{Obtenção de tabelas por agregado via API SIDRA - IBGE}
\usage{
tab_agr(agregado)
}
\arguments{
\item{agregado}{Número do agregado.}
}
\value{
Uma lista contendo dois `data.table`s:
\itemize{
  \item{\code{pesquisas}: Um `data.table` com o ID e o nome das pesquisas relacionadas.}
  \item{\code{tabelas}: Um `data.table` com o ID e o nome das tabelas (agregados) disponíveis para a consulta indicada.}
}
}
\description{
Esta função retorna uma lista com Todas as tabelas para  agregado indicado
}
\examples{
tabs_a70 <- tab_agr('A70')
tab_agr('A70') # imprime os agregados com sua descrição


}
\keyword{IBGE}
\keyword{SIDRA}
\keyword{dados}
\keyword{variáveis}
