% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_agr.R
\name{tab_agrs}
\alias{tab_agrs}
\title{Obtenção de tabelas por agregado via API SIDRA - IBGE}
\usage{
tab_agrs(agregado)
}
\arguments{
\item{agregado}{Número do agregado.}
}
\description{
Esta função retorna uma lista com Todas as tabelas para  agregado indicado
}
\examples{
tabs_a70 <- tab_agrs('A70')
tab_agrs('A70') # imprime os classificadores com sua descrição


}
\keyword{IBGE}
\keyword{SIDRA}
\keyword{dados}
\keyword{variáveis}
