% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_class.R
\name{tab_class}
\alias{tab_class}
\title{Obtenção de dados via API SIDRA - IBGE}
\usage{
tab_class(tabela)
}
\arguments{
\item{tabela}{Número da tabela.}
}
\value{
uma lista de `data.frame`s, onde cada um representa um classificador e suas categorias.
}
\description{
Esta função retorna uma lista com classificadores de uma das tabelas da SIDRA.
}
\examples{
class_ipcaq <- tab_class(1705)
tab_class(1705) # imprime os classificadores com sua descrição
}
\keyword{IBGE}
\keyword{SIDRA}
\keyword{classificadores}
\keyword{dados}
