% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_fonte.R
\name{tab_fonte}
\alias{tab_fonte}
\title{Obtenção de dados via API SIDRA - IBGE}
\usage{
tab_fonte(tabela)
}
\arguments{
\item{tabela}{Número da tabela.}
}
\description{
Esta função retorna a fontede uma das tabelas da SIDRA.
}
\examples{
fonte_pam <- tab_fonte(1612)
tab_fonte(1612) # imprime os classificadores com sua descrição
}
\keyword{IBGE}
\keyword{SIDRA}
\keyword{fonte}
\keyword{metadados}
