% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_meta.R
\name{tab_meta}
\alias{tab_meta}
\title{Obtenção de dados via API SIDRA - IBGE}
\usage{
tab_meta(tabela)
}
\arguments{
\item{tabela}{Número da tabela.}
}
\description{
Esta função retorna uma lista com os ids e o conteúdo da descrição da tabela solicitada.
}
\examples{
meta_ipcaq <- tab_meta(1705)
}
\keyword{IBGE}
\keyword{SIDRA}
\keyword{dados}
\keyword{metadados}
