% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_niveis.R
\name{tab_niveis}
\alias{tab_niveis}
\title{Obtenção de dados via API SIDRA - IBGE}
\usage{
tab_niveis(tabela)
}
\arguments{
\item{tabela}{Número da tabela.}
}
\description{
Esta função retorna uma lista com níveis territoriais disponíveis
de uma das tabelas da SIDRA.
}
\examples{
niveis_ipca15 <- tab_niveis(1705)
tab_niveis(1705) # imprime os classificadores com sua descrição
}
\keyword{IBGE}
\keyword{SIDRA}
\keyword{dados}
\keyword{localidade}
