% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_search.R
\name{tab_search}
\alias{tab_search}
\title{Busca de tabelas/agregados ou variáveis via API SIDRA - IBGE}
\usage{
tab_search(termo)
}
\arguments{
\item{termo}{Termo a buscar.}
}
\value{
um `data.frame` (especificamente um `data.table`) com as colunas:
\itemize{
  \item{\code{id}: um vetor com os ids das tabelas em que termo encontrado}
  \item{\code{literal}: o texto onde ocorreu match}
  \item{\code{agregacao}: codigo string da agregação, cf. `tab_agr()`.}
}
}
\description{
Esta função retorna uma lista com agregados, tabelas ou variáveis da SIDRA
que possuem o termo buscado.
}
\examples{
tabs_ipca <- tab_search('IPCA15')
tab_search('IPCA15') # imprime tabelas/agregados/variáveis com o termo pesquisado.
}
\keyword{IBGE}
\keyword{SIDRA}
\keyword{dados}
\keyword{search}
