% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_vars.R
\name{tab_vars}
\alias{tab_vars}
\title{Obtenção de dados via API SIDRA - IBGE}
\usage{
tab_vars(tabela)
}
\arguments{
\item{tabela}{Número da tabela.}
}
\value{
um `data.frame` (`data.table`) com as seguintes colunas:
\itemize{
  \item{\code{id}: ids dos indicadores}
  \item{\code{nome}: nome dos indicadores}
  \item{\code{unidade}: unidade de medida do indicador.}
  \item{\code{sumarizacao}: tipo de agregação.}
}
}
\description{
Esta função retorna uma lista com variáveis de uma das tabelas da SIDRA.
}
\examples{
vars_ipcaq <- tab_vars(1705)
tab_vars(1705) # imprime os classificadores com sua descrição
}
\keyword{IBGE}
\keyword{SIDRA}
\keyword{dados}
\keyword{variáveis}
