% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_fonte.R
\name{tab_fonte}
\alias{tab_fonte}
\title{Obtenção de dados via API SIDRA - IBGE}
\usage{
tab_fonte(tabela)
}
\arguments{
\item{tabela}{Número da tabela.}
}
\value{
uma `string` (vetor de caracteres de comprimento 1) com o nome da fonte
dos dados da tabela solicitada
}
\description{
Esta função retorna a fonte de uma das tabelas da SIDRA.
}
\examples{
fonte_ipcaq <- tab_fonte(1705)
tab_fonte(1705) # imprime o nome da fonte
}
\keyword{IBGE}
\keyword{SIDRA}
\keyword{fonte}
\keyword{metadados}
