% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_meta.R
\name{tab_meta}
\alias{tab_meta}
\title{Obtenção de dados via API SIDRA - IBGE}
\usage{
tab_meta(tabela)
}
\arguments{
\item{tabela}{Número da tabela.}
}
\value{
Uma `lista` contendo os metadados da tabela solicitada. A lista inclui elementos como:
\itemize{
  \item{\code{nivelTerritorial}: um vetor com os níveis territoriais disponíveis.}
  \item{\code{variaveis}: um `data.frame` com as variáveis da tabela.}
  \item{\code{periodos}: um vetor com os períodos disponíveis.}
  \item{\code{classificacoes}: uma lista de `data.frame`s, onde cada um representa um classificador e suas categorias.}
  \item{Outros metadados diversos da tabela, como nome, pesquisa, assunto, etc.}
}
}
\description{
Esta função retorna uma lista com os ids e o conteúdo da descrição da tabela solicitada.
}
\examples{
meta_ipcaq <- tab_meta(1705)
}
\keyword{IBGE}
\keyword{SIDRA}
\keyword{dados}
\keyword{metadados}
