% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_niveis.R
\name{tab_niveis}
\alias{tab_niveis}
\title{Obtenção de dados via API SIDRA - IBGE}
\usage{
tab_niveis(tabela)
}
\arguments{
\item{tabela}{Número da tabela.}
}
\value{
Um `data.frame` (especificamente, um `data.table`) que lista todas as
 localidades disponíveis para a tabela, detalhando o ID e o nome de cada
 localidade, bem como o ID e o nome do nível geográfico correspondente
 (ex: 'N3' para "Unidade da Federação").
}
\description{
Esta função retorna uma lista com níveis territoriais disponíveis
de uma das tabelas da SIDRA.
}
\examples{
niveis_ipca15 <- tab_niveis(1705)
tab_niveis(1705) # imprime os níveis territoriais da tabela solicitada
}
\keyword{IBGE}
\keyword{SIDRA}
\keyword{dados}
\keyword{localidade}
