% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_periodos.R
\name{tab_periodos}
\alias{tab_periodos}
\title{Obtenção de dados via API SIDRA - IBGE}
\usage{
tab_periodos(tabela)
}
\arguments{
\item{tabela}{Número da tabela.}
}
\value{
um vetor `integer` ou `character` com os períodos disponíveis da tabela solicitada
}
\description{
Esta função retorna uma lista com periodos de uma das tabelas da SIDRA.
}
\examples{
periodos_ipcaq <- tab_periodos(1705)
tab_periodos(1705) # imprime os períodos disponíveis da tabela
}
\keyword{IBGE}
\keyword{SIDRA}
\keyword{dados}
\keyword{periodos}
