% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_periodos.R
\name{tab_periodos}
\alias{tab_periodos}
\title{Obtenção de dados via API SIDRA - IBGE}
\usage{
tab_periodos(tabela)
}
\arguments{
\item{tabela}{Número da tabela.}
}
\description{
Esta função retorna uma lista com periodos de uma das tabelas da SIDRA.
}
\examples{
periodos_pam <- tab_periodos(1612)
tab_periodos(1612) # imprime os classificadores com sua descrição
}
\keyword{IBGE}
\keyword{SIDRA}
\keyword{dados}
\keyword{periodos}
