% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_search.R
\name{tab_search}
\alias{tab_search}
\title{Busca de tabelas/agregados ou variáveis via API SIDRA - IBGE}
\usage{
tab_search(termo)
}
\arguments{
\item{termo}{Termo.}
}
\description{
Esta função retorna uma lista com agregados, tabelas ou variáveis da SIDRA
que possuem o termo buscado.
}
\examples{
class_pam <- tab_class(1612)
tab_class(1612) # imprime os classificadores com sua descrição
}
\keyword{IBGE}
\keyword{SIDRA}
\keyword{dados}
\keyword{search}
