% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_vars.R
\name{tab_vars}
\alias{tab_vars}
\title{Obtenção de dados via API SIDRA - IBGE}
\usage{
tab_vars(tabela)
}
\arguments{
\item{tabela}{Número da tabela.}
}
\description{
Esta função retorna uma lista com variáveis de uma das tabelas da SIDRA.
}
\examples{
vars_pam <- tab_vars(1612)
tab_vars(1612) # imprime os classificadores com sua descrição
}
\keyword{IBGE}
\keyword{SIDRA}
\keyword{dados}
\keyword{variáveis}
